rename query helper file
